Separate classes into respective packages
