--- src/main/java/DwukeException.java
@@ -1,8 +1,0 @@
-/**
- * This class represents exceptions specific to Dwuke.
- */
-public class DwukeException extends Exception {
-    DwukeException(String message) {
-        super("oops!!! " + message);
-    }
-}
