+++ src/main/java/dwuke/Dwuke.java
@@ -1,3 +1,9 @@
+package dwuke;
+
+import dwuke.task.TaskList;
+import dwuke.command.Command;
+import dwuke.command.ExitCommand;
+
 /**
  * Driver class of Dwuke.
  */
