+++ src/main/java/dwuke/command/UnmarkCommand.java
@@ -1,3 +1,9 @@
+package dwuke.command;
+
+import dwuke.task.Task;
+import dwuke.task.TaskList;
+import dwuke.DwukeException;
+
 /**
  * This class encapsulates an unmark command from the user.
  */
@@ -6,7 +12,14 @@ public class UnmarkCommand extends Command {
     private TaskList taskList;
     private int pos;
 
-    UnmarkCommand(TaskList taskList, int pos) throws DwukeException {
+    /**
+     * Creates an UnmarkCommand with the given TaskList and position.
+     *
+     * @param taskList The TaskList.
+     * @param pos The position.
+     * @throws DwukeException If the position is not in the TaskList.
+     */
+    public UnmarkCommand(TaskList taskList, int pos) throws DwukeException {
         if (pos < 0 || pos > taskList.size() - 1) {
             throw new DwukeException("da wist doesn't have a task with da number :3");
         }
